+++ example/lib/views/entry_point.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+import 'package:localize_and_translate/localize_and_translate.dart';
+import 'package:localize_and_translate_example/views/screen1.dart';
+
+class MyApp extends StatefulWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  @override
+  MyAppState createState() => MyAppState();
+}
+
+class MyAppState extends State<MyApp> {
+  @override
+  Widget build(BuildContext context) {
+    debugPrint('${translator.activeLocale}');
+    return MaterialApp(
+      home: const Screen1(),
+      localizationsDelegates: translator.delegates,
+      locale: translator.activeLocale,
+      supportedLocales: translator.locals(),
+    );
+  }
+}