+++ example/lib/views/screen1.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+import 'package:localize_and_translate/localize_and_translate.dart';
+import 'package:localize_and_translate_example/views/screen2.dart';
+
+class Screen1 extends StatefulWidget {
+  const Screen1({Key? key}) : super(key: key);
+
+  @override
+  State<Screen1> createState() => _Screen1State();
+}
+
+class _Screen1State extends State<Screen1> {
+  @override
+  void initState() {
+    debugPrint('app restarted');
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: OutlinedButton(
+            onPressed: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute<Widget>(
+                  builder: (BuildContext builder) => const Screen2(),
+                ),
+              );
+            },
+            child: Text('screen2'.tr())),
+      ),
+    );
+  }
+}