+++ lib/src/extensions.dart
@@ -1,0 +1,5 @@
+import 'package:localize_and_translate/localize_and_translate.dart';
+
+extension Translation on String {
+  String tr() => translator.translate(this);
+}