+++ CHANGELOG.md
@@ -1,32 +1,46 @@
-## Next
+# Next
+
 * Improvements
+* Auto state change after change lang
+
+## 4.1.2-dev
+
+* Re-Structure Package
+* Rebuild after change language .. same screen (state)
 
 ## 4.1.0, 4.1.1
+
 * Remove Google Translate
 * Add String Extension .tr()
 * Enhance Performance
 
 ## 4.0.0, 4.0.1
+
 * Migrate To Null-Safe
 
 ## 3.0.3
+
 * Bug Fix
 * Upgrade Packages
 * Remove intl
 
 ## 3.0.2
+
 * Bug Fix
 * Remove `LIST_OF_LANGS` & `LANGS_DIR`
 * `init()` now takes all needed parameters
 * Google Translate : added with developer-based api key -> NOT YET TESTED
 
 ## 2.2.1
-* `googleTranslate()` : Remove 
+
+* `googleTranslate()` : Remove
 
 ## 2.1.2
+
 * Plugin -> Package
 
 ## 2.0.0 : 2.1.1
+
 * translate() enhance
 * translate() with keys
 * googleTranslate() added
@@ -35,22 +49,25 @@
 * FIXs, docs, video docs
 
 ## 1.2.0 | 1.2.1
+
 * Performance Improvements
 * Docs Improvements
 
 ## 1.1.1+6
+
 * add isDirectionRTL method
 
-
 ## 1.0.1+5
+
 * Documentation Enhancements
 * iOS Delegates
 
 ## 1.0.1
+
 * first lanuch
 * added main methods
 
 ## 0.0.1
+
 * init
 * add description
-